Restructure to include a unit test harness
