+++ pubspec.yaml
@@ -6,3 +6,6 @@ homepage: https://github.com/dart-lang/coverage
 dependencies:
   args: '>=0.9.0<2.0.0'
   path: '>=0.9.0<2.0.0'
+dev_dependencies:
+  unittest: any
+  hop: any
